chore: remove Jest
Justification:

near-runner-jest is now marked as deprecated on NPM, where it links to
the specific commit on GitHub where it still existed. We do not plan to
continue improving this library, so keeping it in our repository just
adds dead weight. It increases our build times and potential for
confusion, and sets expectations that we are actually supporting this
library, even if we'd rather not.

If someone from the community wants to ressurect this code, they can do
so. But I think at this point it makes sense for us to stay
hyper-focused on our premier developer journey, which is
near-runner-ava.