feat: add example Wasm, demonstrate & comment
This sets up the bootstrapped project with an example Wasm file. This
means that the tests will work out the gate, so people can feel more
confident that they know how to change it to get it to work for their
own project.

This also adds a README's worth of comments to the example test file,
since I assume people will mostly read this file instead of a README.