feat: allow broader nodejs range
This was originally restricted when we added linting with XO (#38,
da92f992104b582dd4a096d01e538cd04fa93184), since `@eslint/eslintrc@1.0.1`
requires node version "^12.22.0 || ^14.17.0 || >=16.0.0"

Given that we still use XO in the monorepo, we still need the monorepo
root package.json to require a subset of that range, but we can loosen
the restriction on our individual packages.

This also adds a new script & accompanying GitHub Action to test the
starter project using an expanded NodeJS range.