fix: fix version of near-runner-ava in top-level (#84)
* fix: fix version of near-runner-ava in top-level
all tests and linting were failing because of a mismatch that happened
between different branches

* Publish

In #82 / 7686197894bae2afb314338429d97ba574fddc74 we stopped removing
the test-near-runner-ava directory after the
packages/ava/__tests__/no-install run. We need to ignore this directory
in the main test run. If it's there, concurrent tests will sometimes see
this directory and sometimes not, causing inconsistent builds.