fix: Update length of test accountIds to meet new testnet requirement
The newest roll out of testnet seems to reserve the right to create account Ids that are less than 33 characters long.  Perviously we had generated ones that were 14 as to save characters for sub-accounts.  However, with this requirement we must update to 33, leave 31 characters for sub-accounts.

Also this PR fixes env var for selecting the network.